Complete✅

+++ controllers/contactController.js
@@ -6,15 +6,15 @@ const Contact = require('../models/contact.model');
 
 // @desc Get all contacts
 // @route GET /api/contacts
-// @access public
+// @access private
 const getContact = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({user_id: req.user.id});
   res.status(200).json(contacts);
 });
 
 // @desc Get a contact
 // @route GET /api/contacts/:id
-// @access public
+// @access private
 const getAContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
@@ -26,7 +26,7 @@ const getAContact = asyncHandler(async (req, res) => {
 
 // @desc Creating a contact
 // @route POST /api/contacts
-// @access public
+// @access private
 const createContact = asyncHandler(async (req, res) => {
   console.log(`The request body is: ${req.body}`);
   const { name, email, phone } = req.body;
@@ -38,18 +38,24 @@ const createContact = asyncHandler(async (req, res) => {
     name: name,
     email: email,
     phone: phone,
+    user_id: req.user.id
   });
   res.status(200).json(contact);
 });
 
 // @desc Updating a contacts
 // @route PUT /api/contacts/:id
-// @access public
+// @access private
 const updateContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error('ERROR: Contact Not Found!');
+  }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403)
+    throw new Error(`User don't have permission to update other user contacts!`)
   }
 
   const updatedContact = await Contact.findByIdAndUpdate(
@@ -63,12 +69,16 @@ const updateContact = asyncHandler(async (req, res) => {
 
 // @desc Deleting a contacts
 // @route DELETE /api/contacts/:id
-// @access public
+// @access private
 const deleteContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error('ERROR: Contact Not Found!');
+  }
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403)
+    throw new Error(`User don't have permission to update other user contacts!`)
   }
   await Contact.deleteOne(contact);
   console.log(res.deletedCount)
